perf(workspace): split domain catalog from graph-data; lazy-load picker
Direct follow-up to #240. Previously DomainSelector + DemoFlowPlayer
couldn't be lazy-loaded because four other modules (HeaderBar,
copilot-actions, copilot-context, DemoFlowPlayer itself) imported
named exports from DomainSelector — dragging the whole component +
its transitive graph-data imports into the critical-path bundle.

Refactor
- New `src/lib/domains.ts` — metadata only (DOMAIN_CARDS, DOMAIN_GROUPS,
  PERSONAS, PERSONA_GROUPS, DomainCard, DomainGroup, Persona). No
  graph-data imports. Safe to pull into the initial bundle.
- New `src/lib/build-domain-graph.ts` — owns `buildGraphFromDomains`
  (the only thing that needs MAIN_GRAPH / ATHENA_GRAPH / T1D_GRAPH /
  VX880_GRAPH) plus a `DATASET_NODE_COUNTS` constant for the picker's
  "X nodes will load" preview.
- DomainSelector imports from both new modules. Re-exports the old
  surface (DOMAIN_CARDS, buildGraphFromDomains) for any straggler.
- HeaderBar / copilot-actions / copilot-context / DemoFlowPlayer all
  switched to the canonical new paths so static importers pull only
  the catalog, not the graph data.
- src/app/page.tsx: `DomainSelector` and `DemoFlowPlayerHost` are now
  `next/dynamic({ ssr: false })`. Both render nothing until user-
  triggered, so the deferred chunk fetch is invisible.

Outcome
- 722/722 vitest pass.
- tsc --noEmit clean.
- Critical-path bundle should drop noticeably — DomainSelector chunk
  carries the four graph-data modules (~3000 lines combined) +
  framer-motion + the demo-flow picker. Confirm with
  `ANALYZE=true npm run build` next perf pass.
- Audit-log row appended to docs/PERFORMANCE.md.
