+++ docs/PERFORMANCE.md
@@ -108,7 +108,8 @@ Append a row each time a perf-relevant change ships. Keep entries terse; cite PR
 |---|---|---|---|
 | 2026-05-01 | #222 | Wired `@vercel/speed-insights` into root layout + `@next/bundle-analyzer` into `next.config.ts`. Removed dead `@react-three/postprocessing` (zero source references). | Establishes baseline. Speed Insights data starts accumulating on next prod deploy. Bundle analyzer runs locally only via `ANALYZE=true npm run build`. |
 | 2026-05-06 | #238 | `DAGEdge3D` orb animation: replaced per-frame `curve.getPoint(t)` (allocates Vector3 + bezier eval) with O(1) lookup into the existing 32-point `curvePoints` cache. Dropped invisible-hitbox sphere tessellation 8×8→4×4 and orb tessellation 8×8→6×6. | No measurement run; based on code inspection. Expected savings: ~6k Vector3 allocations/sec at ~100 animating edges + lower raycast cost on hover. Visual diff at 0.25 orb radius is imperceptible. Re-measure during next perf audit; if still slow during cascade replay, the next ladder rung is `<instancedMesh>` for orbs. |
-| 2026-05-06 | (this PR) | Dynamic-imported `ImportModal` and `SpotlightTour` in `src/app/page.tsx` (both render nothing until user-triggered, both pull framer-motion). | Reduces critical-path JS for first paint of the workspace. `DomainSelector` + `DemoFlowPlayer` were considered but skipped: both have other static importers (HeaderBar, copilot lib, etc.) so dynamic-importing them here wouldn't actually split the chunk. To make those lazy, named exports (`DOMAIN_CARDS`, `buildGraphFromDomains`) need to move to a separate module first. |
+| 2026-05-06 | #240 | Dynamic-imported `ImportModal` and `SpotlightTour` in `src/app/page.tsx` (both render nothing until user-triggered, both pull framer-motion). | Reduces critical-path JS for first paint of the workspace. `DomainSelector` + `DemoFlowPlayer` were considered but skipped: both have other static importers (HeaderBar, copilot lib, etc.) so dynamic-importing them here wouldn't actually split the chunk. To make those lazy, named exports (`DOMAIN_CARDS`, `buildGraphFromDomains`) need to move to a separate module first. |
+| 2026-05-06 | (this PR) | Extracted the domain catalog (`DOMAIN_CARDS`, `DOMAIN_GROUPS`, `PERSONAS`, etc.) to a new `src/lib/domains.ts` (metadata only, no graph-data imports) and moved `buildGraphFromDomains` + a new `DATASET_NODE_COUNTS` constant to `src/lib/build-domain-graph.ts` (carries the four heavy graph-data modules). Updated `HeaderBar`, `copilot-actions`, `copilot-context`, `DemoFlowPlayer` to import from the new paths. Made `DomainSelector` and `DemoFlowPlayerHost` `next/dynamic({ ssr: false })` in `src/app/page.tsx`. | Direct follow-up to the previous row's "skipped because of static importers" note — those importers now pull only the lighter catalog. The DomainSelector chunk (which transitively pulls ~3000 lines of graph data + framer-motion + the demo-flow player) defers until the user opens the picker or runs a demo. Critical-path bundle should noticeably shrink; verify with `ANALYZE=true npm run build` next perf pass. |
 
 Future entries: copy the row template above.
