+++ src/app/page.tsx
@@ -13,9 +13,7 @@ import ModulePanel from "@/components/ModulePanel";
 import StructuralMetrics from "@/components/StructuralMetrics";
 import CausalDAG2D from "@/components/CausalDAG2D";
 import TimeDial from "@/components/TimeDial";
-import DomainSelector from "@/components/DomainSelector";
 import FeedbackWidget from "@/components/FeedbackWidget";
-import { DemoFlowPlayerHost } from "@/components/DemoFlowPlayer";
 import TimeSeriesOverlay from "@/components/TimeSeriesOverlay";
 
 // Lazy-loaded modals — neither is on the critical path. Both render
@@ -29,6 +27,25 @@ const ImportModal = dynamic(
 );
 const SpotlightTour = dynamic(
   () => import("@/components/SpotlightTour"),
+  { ssr: false }
+);
+
+// DomainSelector + DemoFlowPlayerHost both transitively pull the
+// four large graph-data modules (~3000 lines combined) via
+// `@/lib/build-domain-graph`. They render nothing until the user
+// opens the picker / runs a demo, so defer their chunks. The catalog
+// surface used elsewhere (HeaderBar's tab labels, copilot context)
+// imports from the lighter `@/lib/domains` module which is unaffected
+// by this lazy load.
+const DomainSelector = dynamic(
+  () => import("@/components/DomainSelector"),
+  { ssr: false }
+);
+const DemoFlowPlayerHost = dynamic(
+  () =>
+    import("@/components/DemoFlowPlayer").then((m) => ({
+      default: m.DemoFlowPlayerHost,
+    })),
   { ssr: false }
 );
